Also adding the link to the tutorial files

+++ mk_links.sh
@@ -3,6 +3,12 @@ HEPTutorialDIR=../HEPTutorial
 
 for object in `echo "MyElectron MyJet MyMuon MyPhoton"`
 do
-	ln -s ${HEPTutorialDIR}/${object}.{C,h} .
-	echo .L ${object}.C+ | root.exe -l -b
+	if [[ ! -f ${object}.C ]] 
+	then
+		ln -s ${HEPTutorialDIR}/${object}.{C,h} .
+		echo .L ${object}.C+ | root.exe -l -b
+	fi
 done
+
+ln -s -d ${HEPTutorialDIR}/files .
+
